backend/src/api/requirements.txt
fastapi>=0.104.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
uvicorn>=0.23.2
httpx>=0.24.1
python-jose>=3.3.0
python-multipart>=0.0.6
Jinja2>=3.1.2
authlib>=1.2.1
boto3>=1.28.63
mangum>=0.17.0
